Update sbt: the old sbt crashes on modern Java
On a candidate's machine with Java 15, the old version of sbt (1.2) did not seem able to handle the Java 9 module system, so far as I could see. Just upgrading to the latest version of sbt (1.5.2) fixed the problem for them.

I've not been able to reproduce the problem with old sbt 1.2 on my own laptop - this might be because I'm using the latest sbt launcher script.